* commit 'e0f2e575c5f18eef15d8c14ed4fd8f8ecdaf857a': (268 commits)
  ENH: Set a larger gutter on the matrix.
  ENH: Added API to store, set and retrieve the active plot.
  ENH: Added chart spans to the chart matrix.
  KWSys Nightly Date Stamp
  ENH: Add a placeholder for the active plot.
  ENH: Synchronize axis ranges in the matrix.
  ENH: Added a setter that uses double* for axis range.
  ENH: Removed the Range ivar from vtkAxis.
  ENH: Add event to chart/axis when range is changed.
  ENH: Use the gutter and border in chart matrix.
  ENH: Added code for axes to rectangle in charts.
  ENH: Changed the chart layout strategy to an enum.
  ENH: Added scatter plot matrix column visibility.
  KWSys Nightly Date Stamp
  ENH: Added simple histograms to the matrix.
  BUG: Fixed Common header testing.
  ENH: Added a fuzzy comparison function.
  ENH: Ensure charts are removed correctly in resizes.
  ENH: Just fill the lower triangle with charts.
  KWSys Nightly Date Stamp
  ...

Conflicts:
	Charts/vtkAxis.cxx

+++ Rendering/vtkFreeTypeTools.cxx
@@ -222,7 +222,7 @@ vtkFreeTypeToolsFaceRequester(FTC_FaceID face_id,
   // Map the ID to a text property
   vtkSmartPointer<vtkTextProperty> tprop =
       vtkSmartPointer<vtkTextProperty>::New();
-  self->MapIdToTextProperty(reinterpret_cast<unsigned long>(face_id), tprop);
+  self->MapIdToTextProperty(reinterpret_cast<intptr_t>(face_id), tprop);
 
   // Fonts, organized by [Family][Bold][Italic]
   static EmbeddedFontStruct EmbeddedFonts[3][2][2] =
